requirements.txt
matplotlib
requests
python-dotenv
pytest
allure-pytest
pytest-mock
